# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clients/upsertClientUser": {
            "post": {
                "operationId": "adf25868edf5a07747f0195c223f0591",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "clientcode",
                                    "usercode"
                                ],
                                "properties": {
                                    "clientcode": {
                                        "description": "Client Code",
                                        "type": "string"
                                    },
                                    "usercode": {
                                        "description": "User Code",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "User Name",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User Password, will be encrypted and saved",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add/update client user"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    }
}